+++ www/app/client/tasks/TestTask.class.php
@@ -2,9 +2,6 @@
 
 namespace tasks;
 
-use herosphp\db\DBFactory;
-use herosphp\db\entity\MongoEntity;
-use herosphp\lock\SynLockFactory;
 use tasks\interfaces\ITask;
 use herosphp\core\Loader;
 
@@ -18,23 +15,7 @@ class TestTask implements ITask {
 
         public function run() {
 
-//            $lock = SynLockFactory::getFileSynLock(0x1234);
-//            tprintError("try to get the lock....");
-//            $lock->tryLock();
-//            tprintOk("get the lock.");
-//            sleep(10);
-//            tprintWarning("release the lock.");
-//            $lock->unlock();
-
-
-            $model = Loader::model("news");
-            $timer = timer();
-            for ( $i = 0; $i < 1000000; $i++ ) {
-                $data = array('name' => 'xiaoming', 'pass' => 'xiaoming_pass', 'address' => 'china ShenZhen');
-                $model->insert($data);
-            }
-
-            tprintOk("插入完成，耗时：".(timer() - $timer)." 秒");
+            tprintOk("Hello, Body! You are running a Task.");
         }
 
 }